[storage-if] move DbReader to storage/storage-interface
