+++ secure/key-manager/src/tests.rs
@@ -21,10 +21,11 @@ use libra_types::{
     validator_set::ValidatorSetResource,
 };
 use libra_vm::LibraVM;
-use libradb::{DbReader, LibraDB};
+use libradb::LibraDB;
 use rand::{rngs::StdRng, SeedableRng};
 use std::{cell::RefCell, collections::BTreeMap, convert::TryFrom, sync::Arc, time::Duration};
 use storage_client::{StorageReadServiceClient, StorageWriteServiceClient};
+use storage_interface::DbReader;
 use tokio::runtime::Runtime;
 
 struct Node {
