+++ storage/libradb/Cargo.toml
@@ -31,6 +31,7 @@ jellyfish-merkle = { path = "../jellyfish-merkle", version = "0.1.0" }
 libra-logger = { path = "../../common/logger", version = "0.1.0" }
 libra-metrics = { path = "../../common/metrics", version = "0.1.0" }
 schemadb = { path = "../schemadb", version = "0.1.0" }
+storage-interface = { path = "../storage-interface", version = "0.1.0" }
 storage-proto = { path = "../storage-proto", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
 libra-proptest-helpers = { path = "../../common/proptest-helpers", version = "0.1.0", optional = true }
