+++ storage/storage-interface/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "libra-storage-inspector"
+name = "storage-interface"
 version = "0.1.0"
 authors = ["Libra Association <opensource@libra.org>"]
+description = "Libra storage interface"
 repository = "https://github.com/libra/libra"
 homepage = "https://libra.org"
 license = "Apache-2.0"
@@ -10,12 +11,7 @@ edition = "2018"
 
 [dependencies]
 anyhow = "1.0"
-structopt = "0.3.12"
-tempfile = "3.1.0"
 
-libradb = { path = "../libradb", version = "0.1.0" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
+storage-proto = { path = "../storage-proto", version = "0.1.0" }
 libra-types = { path = "../../types", version = "0.1.0" }
-libra-logger = { path = "../../common/logger", version = "0.1.0" }
-storage-interface = { path = "../storage-interface", version = "0.1.0" }
-transaction-builder = { path = "../../language/transaction-builder", version = "0.1.0" }
